requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.1.0
numpy>=1.24.0
websocket-client>=1.6.0
python-dotenv>=1.0.0